[inductor] Move test_inductor_sequence_nr out of test_aot_inductor
Summary: The initial PR that added test_inductor_sequence_nr (https://github.com/pytorch/pytorch/pull/103129) seems to think test_aot_inductor is to test aot_autograd + inductor. Move the test to test_torchinductor instead, which can simplify the logic for test_aot_inductor.

[ghstack-poisoned]
